[Style] onboardingstartview 위치 수정
